Export type `GreaterThan`, `GreaterThanOrEqual`, `LessThan`, `LessThanOrEqual`
