+++ source/add.d.ts
@@ -1,0 +1,66 @@
+import type {NumberAbsolute, BuildTuple, And, Or, ArrayMax, ArrayMin} from './internal';
+import type {IsEqual} from './is-equal';
+import type {PositiveInfinity, NegativeInfinity, IsNegative} from './numeric';
+import type {Subtract} from './subtract';
+
+/**
+Returns the sum of the two input numbers.
+
+Note:
+- A or B can only support `-999` ~ `999`.
+- A and B can only be small integers, less than 1000.
+- If the result is negative, you can only get `number`.
+
+@example
+```
+import type {Add} from 'type-fest';
+
+Add<111, 222>;
+//=> 333
+
+Add<-111, 222>;
+//=> 111
+
+Add<111, -222>;
+//=> number
+
+Add<PositiveInfinity, -9999>;
+//=> PositiveInfinity
+
+Add<PositiveInfinity, NegativeInfinity>;
+//=> number
+```
+
+@category Numeric
+*/
+// TODO: Support big integer and negative number.
+export type Add<A extends number, B extends number> = [
+	IsEqual<A, PositiveInfinity>, IsEqual<A, NegativeInfinity>,
+	IsEqual<B, PositiveInfinity>, IsEqual<B, NegativeInfinity>,
+] extends infer R extends [boolean, boolean, boolean, boolean]
+	? Or<
+	And<IsEqual<R[0], true>, IsEqual<R[3], false>>,
+	And<IsEqual<R[2], true>, IsEqual<R[1], false>>
+	> extends true
+		? PositiveInfinity
+		: Or<
+		And<IsEqual<R[1], true>, IsEqual<R[2], false>>,
+		And<IsEqual<R[3], true>, IsEqual<R[0], false>>
+		> extends true
+			? NegativeInfinity
+			: true extends R[number]
+				? number
+				: ([IsNegative<A>, IsNegative<B>] extends infer R
+					? [false, false] extends R
+						? [...BuildTuple<A>, ...BuildTuple<B>]['length']
+						: [true, true] extends R
+							? number
+							: ArrayMax<[NumberAbsolute<A>, NumberAbsolute<B>]> extends infer Max_
+								? ArrayMin<[NumberAbsolute<A>, NumberAbsolute<B>]> extends infer Min_ extends number
+									? Max_ extends A | B
+										? Subtract<Max_, Min_>
+										: number
+									: never
+								: never
+					: never) & number
+	: never;
