+++ source/greater-than.d.ts
@@ -3,7 +3,7 @@ import type {IsEqual} from './is-equal';
 import type {PositiveInfinity, NegativeInfinity, IsNegative} from './numeric';
 
 /**
-Returns a boolean for whether a given number is greater than another number.
+Returns a boolean for whether A is greater than B.
 
 @example
 ```
